move things into the tests directory
