+++ tests/e2e/settings-profile.test.ts
@@ -1,6 +1,6 @@
 import { faker } from '@faker-js/faker'
 import { expect, insertNewUser, test } from '../playwright-utils.ts'
-import { createUser } from '../../prisma/seed-utils.ts'
+import { createUser } from '../../tests/db-utils.ts'
 import { verifyLogin } from '~/utils/auth.server.ts'
 
 test('Users can update their basic info', async ({ login, page }) => {
