+++ tests/mocks/utils.ts
@@ -4,7 +4,7 @@ import { fileURLToPath } from 'url'
 import { z } from 'zod'
 
 const __dirname = path.dirname(fileURLToPath(import.meta.url))
-const fixturesDirPath = path.join(__dirname, `./fixtures`)
+const fixturesDirPath = path.join(__dirname, '..', 'fixtures')
 
 export async function readFixture(subdir: string, name: string) {
 	return fsExtra.readJSON(path.join(fixturesDirPath, subdir, `${name}.json`))
