+++ tests/playwright-utils.ts
@@ -3,13 +3,11 @@ import { parse } from 'cookie'
 import { authenticator, getPasswordHash } from '~/utils/auth.server.ts'
 import { prisma } from '~/utils/db.server.ts'
 import { commitSession, getSession } from '~/utils/session.server.ts'
-import { createUser } from '../prisma/seed-utils.ts'
+import { createUser } from '../tests/db-utils.ts'
 
 export const dataCleanup = {
 	users: new Set<string>(),
 }
-
-export { readEmail } from '../mocks/utils.ts'
 
 export function deleteUserByUsername(username: string) {
 	return prisma.user.delete({ where: { username } })
