+++ tests/setup/vitejs-plugin-react.cjs
