+++ vitest.config.ts
@@ -1,7 +1,7 @@
 /// <reference types="vitest" />
 /// <reference types="vite/client" />
 
-import { react } from './other/test-setup/vitejs-plugin-react.cjs'
+import { react } from './tests/setup/vitejs-plugin-react.cjs'
 import { defineConfig } from 'vite'
 import tsconfigPaths from 'vite-tsconfig-paths'
 
@@ -10,8 +10,8 @@ export default defineConfig({
 	test: {
 		include: ['./app/**/*.test.{ts,tsx}'],
 		environment: 'jsdom',
-		setupFiles: ['./other/test-setup/setup-test-env.ts'],
-		globalSetup: ['./other/test-setup/global-setup.ts'],
+		setupFiles: ['./tests/setup/setup-test-env.ts'],
+		globalSetup: ['./tests/setup/global-setup.ts'],
 		coverage: {
 			include: ['app/**/*.{ts,tsx}'],
 			all: true,
